+++ src/apps/kyberswap-elastic/contracts/ethers/index.ts
@@ -2,14 +2,14 @@
 /* tslint:disable */
 /* eslint-disable */
 export type { Factory } from './Factory';
-export type { Masterchef } from './Masterchef';
+export type { KyberswapElasticLm } from './KyberswapElasticLm';
 export type { Pool } from './Pool';
 export type { PositionManager } from './PositionManager';
 export type { RewardLookerV2 } from './RewardLookerV2';
 export type { TickReader } from './TickReader';
 export * as factories from './factories';
 export { Factory__factory } from './factories/Factory__factory';
-export { Masterchef__factory } from './factories/Masterchef__factory';
+export { KyberswapElasticLm__factory } from './factories/KyberswapElasticLm__factory';
 export { Pool__factory } from './factories/Pool__factory';
 export { PositionManager__factory } from './factories/PositionManager__factory';
 export { RewardLookerV2__factory } from './factories/RewardLookerV2__factory';