+++ src/apps/sushiswap-kashi/ethereum/sushiswap-kashi.leverage.contract-position-fetcher.ts
@@ -1,0 +1,8 @@
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
+import { SushiswapKashiLeverageContractPositionFetcher } from '../common/sushiswap-kashi.leverage.contract-position-fetcher';
+
+@PositionTemplate()
+export class EthereumSushiswapKashiLeverageContractPositionFetcher extends SushiswapKashiLeverageContractPositionFetcher {
+  groupLabel = 'Lending';
+}